+++ AgendaServer/src/main/kotlin/model/User.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.sql.Table
 import org.joda.time.DateTime
 
 object TB_User : Table(){
-    val id_user = integer("id_user").primaryKey().autoIncrement()
+    val id_user = long("id_user").primaryKey().autoIncrement()
     val username = varchar("username", 16)
     val email = varchar("email", 255)
     val password = varchar("password", 32)
@@ -12,7 +12,7 @@ object TB_User : Table(){
 }
 
 data class User(
-        val id_user : Int,
+        val id_user : Long,
         val username : String,
         val email : String,
         val password : String,
@@ -20,7 +20,7 @@ data class User(
 )
 
 data class NewUser(
-        val id_user : Int?,
+        val id_user : Long?,
         val username : String,
         val email : String,
         val password : String,
